% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avail.R
\name{exif_tags}
\alias{exif_tags}
\title{Return a data frame of all possible Exif tags}
\usage{
exif_tags(filter_key = NULL, filter_type = NULL)
}
\arguments{
\item{filter_key}{a regular expression to filter key names by. If \code{NULL} then
no key name filtering will be performed.}

\item{filter_type}{a vector of "ascii", "long", "short", "rational", "byte",
"sshort", "srational", "float", "double", or "comment".
if \code{NULL} then all types will be considered.}
}
\value{
data frame
}
\description{
Return a data frame of all possible Exif tags
}
