% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-exif.R
\name{read_exif}
\alias{read_exif}
\title{Retrieve Exif Data From An Image}
\usage{
read_exif(x)
}
\arguments{
\item{x}{path to image file}
}
\value{
data frame or \code{NULL} if no Exif data is present
}
\description{
Retrieve Exif Data From An Image
}
