% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set-exif.R
\name{set_exif}
\alias{set_exif}
\title{Set Exif tag data on an image  file}
\usage{
set_exif(x, tag, val)
}
\arguments{
\item{x}{path to image file}

\item{tag}{Exif tag name to set (full \code{Title.Case.Dotted} name)}

\item{val}{Exif tag value; \code{rational} should be a two element numeric vector}
}
\value{
data frame of image Exif values
}
\description{
Set metadata entries on image files. Support is planned for the components of
\href{http://www.exiv2.org/metadata.html}{these reference tables} that are able to be
crafted from within R.
}
\details{
Currently only handles:
\itemize{
\item \code{ascii}
\item \code{comment}
\item \code{long}
\item \code{short}
\item \code{srational}
\item \code{rational}
}
}
