% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-exif.R
\name{read_exif}
\alias{read_exif}
\title{Retrieve Exif data from an image file}
\usage{
read_exif(x)
}
\arguments{
\item{x}{path to image file}
}
\value{
data frame or \code{NULL} if no Exif data is present. This does mean that all the
values are conv are \code{toString()}-ified on the C++ side. This could easily return
a named list of real values. Not sure how to handle it, but there are also
other pkgs for handing this.
}
\description{
FEEDBACK WELCOME/ENCOURAGED FOR WHAT THIS SHOULD RETURN
}
