% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set-exif.R
\name{set_exif}
\alias{set_exif}
\title{Set Exif tag data on an image/media file}
\usage{
set_exif(x, tag, val)
}
\arguments{
\item{x}{path to image file}

\item{tag}{Exif tag name to set (full \code{Title.Case.Dotted} name)}

\item{val}{Exif tag value; \code{rational} should be a two element numeric vector}
}
\value{
data frame of image Exif values
}
\description{
CURRENTLY ONLY HANDLES ASCII
}
